Move `deposit-refund.ts` to the `./services/deposits` feature module
This code is part of the user-facing deposit flow so should live in the
`./services/deposits` module.
